Update line-colors.csv
Swapping the source for the Mitteldeutsche Regiobahn with the source for the "VMS-MRB".

As part of the uniform nomenclature, the colours of the operator are to be used for cross-network lines.

Addition of the RE6 and RB 110 lines. 
RB83 (formerly Freiberger Eisenbahn) is not easily possible and has been omitted.