Add colors for VR services in Finland
This uses the green VR color for services that extend to outside the HSL
region and use the green-colored VR rolling stock, while [HSL]'s purple
commuter train color is used for HSL region services that VR is
contracted to operate using HSL-colored and branded rolling stock. A
similar divide is used on e.g. Wikipedia where a single color is needed
for a line.

[HSL]: https://en.wikipedia.org/wiki/Helsinki_Regional_Transport_Authority